requirements.txt
gradio==3.26.0
python-dotenv
litellm>=1.0.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.20
alembic>=1.12.0
pydantic>=2.0.0
pandas>=2.0.0
PyPDF2
python-docx>=0.8.11
pdfplumber>=0.9.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
openai>=1.0.0
anthropic>=0.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
uvicorn[standard]>=0.23.0
